+++ NEWS.md
@@ -1,3 +1,10 @@
+# cansim 0.3.10
+## Minor changes
+* Better error handling when StatCan returns empty tables
+* Add Hierachy for Geography in sqlite tables
+* Better fallback and warning messages when StatCan table categories are internally inconsistent
+* Performance improvements
+
 # cansim 0.3.9
 ## Major changes
 * deprecate `list_cansim_tables` and `serach_cansim_tables` and fallback to corresponding "_cube" methods as Open Data Canada API has changed and similar functionality is available through the "_cube" methods that tie directly into StatCan APIS